refactor: Consolidate stablecoin-symbol tests under api::test
api.rs declares `pub(crate) mod test;` pointing at `api/test.rs`, but
c3ac60d (which moved `record_stablecoin_symbol_rates_received` next to
its caller) added a second inline `#[cfg(test)] mod stablecoin_symbol_metrics`
at the bottom of `api.rs` — leaving the file with two test sources and
a reader asking which one to grep first.

Fold the inline module into `api/test.rs` as a sub-module so all
api-module tests sit under a single `mod test` umbrella, matching the
file's pre-existing convention. The sub-module reaches up to the
parent module's private items via `super::super::` because the function
under test is private to `api.rs`; this is the same visibility trick
`api/test.rs` already uses for `get_exchange_rate_internal` and the
other private helpers it exercises.

Test paths shift from `api::stablecoin_symbol_metrics::*` to
`api::test::stablecoin_symbol_metrics::*` accordingly.
